requirements.txt
numpy 
pillow 
typing 
opencv-python
